requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.0,<6.0
celery>=5.2,<6.0
redis>=4.5,<5.0
gunicorn>=20.1,<21.0
drf-yasg>=1.21,<2.0
django
djangorestframework
django-cors-headers
psycopg2-binary  
Pillow